Fix use after free when node and it's child are queued for destroy
Previosuly, if a node was queued for destroy and one of it's children
nodes was also queued later, the parent node would be destroyed along
with it's children, and afterwards the queued child node would be
accessed to be destroyed, causing a use after free. This commit fixes
that issue by introducing a scaffold_dequeue_destroy function, which
removes a node from the destroy queue. This function is called whenever
a node is destroyed inside of scaffold_node_destroy, guaranteeing that a
node that was destroyed by any means will not be destroyed again when
iterating through the destroy queue.